Feat: Progreso de carrito de compras, orden de los modelos con pascalcase

+++ src/main/java/com/spring/team1/tiendamia/Models/Carrito/CarritoDetalle.java
@@ -1,7 +1,7 @@
 package com.spring.team1.tiendamia.Models.Carrito;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.spring.team1.tiendamia.Models.productos.Variaciones_Producto;
+import com.spring.team1.tiendamia.Models.Productos.Variaciones_Producto;
 
 import jakarta.persistence.*;
 import lombok.Data;
@@ -13,7 +13,7 @@ public class CarritoDetalle {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "idcarrito", nullable = false)
